fix: enforce SA1402 one type per file guideline
