+++ src/FastCharts.Core/DataBinding/Series/MarkerShape.cs
@@ -1,0 +1,43 @@
+namespace FastCharts.Core.DataBinding.Series
+{
+    /// <summary>
+    /// Marker shapes for line series points
+    /// </summary>
+    public enum MarkerShape
+    {
+        /// <summary>
+        /// No marker
+        /// </summary>
+        None,
+
+        /// <summary>
+        /// Circular marker
+        /// </summary>
+        Circle,
+
+        /// <summary>
+        /// Square marker
+        /// </summary>
+        Square,
+
+        /// <summary>
+        /// Diamond marker
+        /// </summary>
+        Diamond,
+
+        /// <summary>
+        /// Triangle marker
+        /// </summary>
+        Triangle,
+
+        /// <summary>
+        /// Cross marker
+        /// </summary>
+        Cross,
+
+        /// <summary>
+        /// Plus marker
+        /// </summary>
+        Plus
+    }
+}
