+++ src/FastCharts.Core/DataBinding/Series/ObservableLineSeries.cs
@@ -94,45 +94,4 @@ namespace FastCharts.Core.DataBinding.Series
             return $"ObservableLineSeries: {_data.Count} points from {sourceType} (X: {XPath}, Y: {YPath})";
         }
     }
-
-    /// <summary>
-    /// Marker shapes for line series points
-    /// </summary>
-    public enum MarkerShape
-    {
-        /// <summary>
-        /// No marker
-        /// </summary>
-        None,
-
-        /// <summary>
-        /// Circular marker
-        /// </summary>
-        Circle,
-
-        /// <summary>
-        /// Square marker
-        /// </summary>
-        Square,
-
-        /// <summary>
-        /// Diamond marker
-        /// </summary>
-        Diamond,
-
-        /// <summary>
-        /// Triangle marker
-        /// </summary>
-        Triangle,
-
-        /// <summary>
-        /// Cross marker
-        /// </summary>
-        Cross,
-
-        /// <summary>
-        /// Plus marker
-        /// </summary>
-        Plus
-    }
 }
