+++ tests/FastCharts.Core.Tests/DataBinding/CategoryValue.cs
@@ -1,0 +1,11 @@
+namespace FastCharts.Core.Tests.DataBinding
+{
+    /// <summary>
+    /// Test data class for category-value pairs
+    /// </summary>
+    public class CategoryValue
+    {
+        public string Category { get; set; } = "";
+        public double Value { get; set; }
+    }
+}
