+++ tests/FastCharts.Core.Tests/DataBinding/ComplexDataPoint.cs
@@ -1,0 +1,11 @@
+namespace FastCharts.Core.Tests.DataBinding
+{
+    /// <summary>
+    /// Test data class for complex data points with nested properties
+    /// </summary>
+    public class ComplexDataPoint
+    {
+        public Point2D Location { get; set; } = new();
+        public int Value { get; set; }
+    }
+}
