+++ tests/FastCharts.Core.Tests/DataBinding/Point2D.cs
@@ -1,0 +1,11 @@
+namespace FastCharts.Core.Tests.DataBinding
+{
+    /// <summary>
+    /// Test data class for 2D point coordinates
+    /// </summary>
+    public class Point2D
+    {
+        public double X { get; set; }
+        public double Y { get; set; }
+    }
+}
